edited spic-rpi.cpp, functions init, deinit and transfer

+++ src/framework/raspberrypi/pal/spic-rpi.hpp
@@ -54,7 +54,7 @@ class SPICRpi: virtual public SPIC
 		uint8_t     mode;
 		uint8_t     clock;
 		uint8_t		channel;
-		uint8_t		speed; 
+		uint32_t	speed; 
 
 	public:
 					SPICRpi();
